+++ sdk/core/core-lro/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.0.1 (Unreleased)
 
+- getOperationState() now returns TState.
+- PollerLike's TState now extends PollOperationState.
+- Since getOperationState() can be overwritten, this is now used to separate between public and private state.
+
 ## 1.0.0 (2019-10-29)
 
 This release marks the general availability of the `@azure/core-lro` package.